+++ base_aux/aux_attr/m2_annot3_cls_keys_as_values.py
@@ -1,6 +1,6 @@
 from typing import *
 
-from .m1_annot_attr1_aux import AnnotAttrAux
+from .m1_annot_attr1_aux import AnnotsAllAux
 
 
 # =====================================================================================================================
@@ -10,7 +10,7 @@ class Meta_AnnotClsKeysAsValues(type):
     if no corresponding annotation - raise!
     """
     def __getattr__(cls, item: str) -> str | NoReturn:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         if item in annots:
             return item
         else:
@@ -18,27 +18,27 @@ class Meta_AnnotClsKeysAsValues(type):
             raise AttributeError(msg)
 
     def __iter__(cls) -> Iterable[str]:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         yield from annots
 
     def __len__(cls) -> int:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         return len(annots)
 
     def __contains__(cls, item: str) -> bool:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         return item in annots
 
     def __getitem__(cls, item: int) -> str | NoReturn:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         return list(annots)[item]
 
     def __str__(cls) -> str:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         return str(tuple(annots))
 
     def __repr__(cls) -> str:
-        annots = AnnotAttrAux(cls).dump_dict__annot_types()
+        annots = AnnotsAllAux(cls).dump_dict__annot_types()
         return f"{cls.__name__}{tuple(annots)}"