+++ base_aux/aux_dict/m3_dict_attr1_simple.py
@@ -170,7 +170,7 @@ class DictAttrAnnotRequired(DictAttr, NestInit_AnnotsRequired):
 
     def check_all_defined_or_raise(self) -> None | NoReturn:
         not_def_list = []
-        nested = AnnotAttrAux(self).dump_dict__annot_types()
+        nested = AnnotsAllAux(self).dump_dict__annot_types()
         for key in nested:
             if key not in self:
                 not_def_list.append(key)