+++ base_aux/base_nest_dunders/m1_init2_annots3_required.py
@@ -14,7 +14,7 @@ class NestInit_AnnotsRequired(NestGAI_AnnotAttrIC):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-        AnnotAttrAux(self).annots__check_all_defined_or_raise()  # check only after superInit!
+        AnnotsAllAux(self).annots__check_all_defined_or_raise()  # check only after superInit!
 
     # TODO: deside is it really need NamedTuple and dataclasses??? seems its not need!!! - NEED!!! realise later!!!